Đưa file OrderController vào thư mục orders Đưa file SaveVoucherController và VoucherController vào thư mục vouchers Sửa thẻ link thêm chọn thêm sản phẩm và mua hàng trong trang giỏ hàng
