+++ src/main/java/vn/edu/hcmuaf/fit/controller/web/orders/OrderController.java
@@ -1,14 +1,15 @@
-package vn.edu.hcmuaf.fit.controller.web;
+package vn.edu.hcmuaf.fit.controller.web.orders;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
 import javax.servlet.annotation.*;
 import java.io.IOException;
 
-@WebServlet(name = "saveVoucher", value = "/saveVoucher")
-public class SaveVoucherController extends HttpServlet {
+@WebServlet(name = "order", value = "/order")
+public class OrderController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        request.getRequestDispatcher("/views/web/order.jsp").forward(request, response);
 
     }
